--- static/login.css
@@ -1,294 +1,0 @@
-*, *::after, *::before {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    user-select: none;
-  }
-  
-  /* Generic */
-  body {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-family: "Montserrat", sans-serif;
-    font-size: 12px;
-    background-color: #ecf0f3;
-    color: #a0a5a8;
-  }
-  
-  /**/
-  .main {
-    position: relative;
-    width: 1000px;
-    min-width: 1000px;
-    min-height: 600px;
-    height: 600px;
-    padding: 25px;
-    background-color: #ecf0f3;
-    box-shadow: 10px 10px 10px #d1d9e6, -10px -10px 10px #f9f9f9;
-    border-radius: 12px;
-    overflow: hidden;
-  }
-  @media (max-width: 1200px) {
-    .main {
-      transform: scale(0.7);
-    }
-  }
-  @media (max-width: 1000px) {
-    .main {
-      transform: scale(0.6);
-    }
-  }
-  @media (max-width: 800px) {
-    .main {
-      transform: scale(0.5);
-    }
-  }
-  @media (max-width: 600px) {
-    .main {
-      transform: scale(0.4);
-    }
-  }
-  
-  .container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    top: 0;
-    width: 600px;
-    height: 100%;
-    padding: 25px;
-    background-color: #ecf0f3;
-    transition: 1.25s;
-  }
-  
-  .form {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  }
-  .form__icon {
-    object-fit: contain;
-    width: 30px;
-    margin: 0 5px;
-    opacity: 0.5;
-    transition: 0.15s;
-  }
-  .form__icon:hover {
-    opacity: 1;
-    transition: 0.15s;
-    cursor: pointer;
-  }
-  .form__input {
-    width: 350px;
-    height: 40px;
-    margin: 4px 0;
-    padding-left: 25px;
-    font-size: 13px;
-    letter-spacing: 0.15px;
-    border: none;
-    outline: none;
-    font-family: "Montserrat", sans-serif;
-    background-color: #ecf0f3;
-    transition: 0.25s ease;
-    border-radius: 8px;
-    box-shadow: inset 2px 2px 4px #d1d9e6, inset -2px -2px 4px #f9f9f9;
-  }
-  .form__input:focus {
-    box-shadow: inset 4px 4px 4px #d1d9e6, inset -4px -4px 4px #f9f9f9;
-  }
-  .form__span {
-    margin-top: 30px;
-    margin-bottom: 12px;
-  }
-  .form__link {
-    color: #181818;
-    font-size: 15px;
-    margin-top: 25px;
-    border-bottom: 1px solid #a0a5a8;
-    line-height: 2;
-  }
-  
-  .title {
-    font-size: 34px;
-    font-weight: 700;
-    line-height: 3;
-    color: #181818;
-  }
-  
-  .description {
-    font-size: 14px;
-    letter-spacing: 0.25px;
-    text-align: center;
-    line-height: 1.6;
-  }
-  
-  .button {
-    width: 180px;
-    height: 50px;
-    border-radius: 25px;
-    margin-top: 50px;
-    font-weight: 700;
-    font-size: 14px;
-    letter-spacing: 1.15px;
-    background-color: #4B70E2;
-    color: #f9f9f9;
-    box-shadow: 8px 8px 16px #d1d9e6, -8px -8px 16px #f9f9f9;
-    border: none;
-    outline: none;
-  }
-  
-  /**/
-  .a-container {
-    z-index: 100;
-    left: calc(100% - 600px );
-  }
-  
-  .b-container {
-    left: calc(100% - 600px );
-    z-index: 0;
-  }
-  
-  .switch {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 400px;
-    padding: 50px;
-    z-index: 200;
-    transition: 1.25s;
-    background-color: #ecf0f3;
-    overflow: hidden;
-    box-shadow: 4px 4px 10px #d1d9e6, -4px -4px 10px #f9f9f9;
-  }
-  .switch__circle {
-    position: absolute;
-    width: 500px;
-    height: 500px;
-    border-radius: 50%;
-    background-color: #ecf0f3;
-    box-shadow: inset 8px 8px 12px #d1d9e6, inset -8px -8px 12px #f9f9f9;
-    bottom: -60%;
-    left: -60%;
-    transition: 1.25s;
-  }
-  .switch__circle--t {
-    top: -30%;
-    left: 60%;
-    width: 300px;
-    height: 300px;
-  }
-  .switch__container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    position: absolute;
-    width: 400px;
-    padding: 50px 55px;
-    transition: 1.25s;
-  }
-  .switch__button {
-    cursor: pointer;
-  }
-  .switch__button:hover {
-    box-shadow: 6px 6px 10px #d1d9e6, -6px -6px 10px #f9f9f9;
-    transform: scale(0.985);
-    transition: 0.25s;
-  }
-  .switch__button:active, .switch__button:focus {
-    box-shadow: 2px 2px 6px #d1d9e6, -2px -2px 6px #f9f9f9;
-    transform: scale(0.97);
-    transition: 0.25s;
-  }
-  
-  /**/
-  .is-txr {
-    left: calc(100% - 400px );
-    transition: 1.25s;
-    transform-origin: left;
-  }
-  
-  .is-txl {
-    left: 0;
-    transition: 1.25s;
-    transform-origin: right;
-  }
-  
-  .is-z200 {
-    z-index: 200;
-    transition: 1.25s;
-  }
-  
-  .is-hidden {
-    visibility: hidden;
-    opacity: 0;
-    position: absolute;
-    transition: 1.25s;
-  }
-  
-  .is-gx {
-    animation: is-gx 1.25s;
-  }
-  
-  @keyframes is-gx {
-    0%, 10%, 100% {
-      width: 400px;
-    }
-    30%, 50% {
-      width: 500px;
-    }
-  }
-
-
-  .email-container {
-    display: flex;
-    align-items: center;
-    width: 350px; /* 같은 너비로 맞추기 */
-    margin: 4px 0; /* 입력 필드와 같은 마진 적용 */
-}
-
-.email-container .form__input {
-    flex: 1; /* 남은 공간을 채움 */
-    margin-right: 10px; /* 버튼과의 간격 */
-}
-
-#check-email {
-    width: 80px; /* 버튼의 너비를 고정 */
-    height: 40px; /* 버튼의 높이를 고정 */
-    border-radius: 5px; /* 둥근 모서리 */
-    margin-top: 0; /* 버튼의 상단 여백을 제거 */
-    /* font-weight: 700; 글자 두껍게 */
-    font-size: 13px; /* 글자 크기 */
-    letter-spacing: 1.15px; /* 글자 간격 조정 */
-    background-color: #ecf0f3; /* 버튼 배경색 */
-    color: #878b8e; /* 글자 색상 */
-    box-shadow: 8px 8px 16px #d1d9e6, -8px -8px 16px #f9f9f9; /* 그림자 효과 */
-    border: #acaeaf solid 2px; /* 기본 테두리 제거 */
-    outline: none; /* 포커스 아웃라인 제거 */
-    cursor: pointer; /* 마우스 포인터 변경 */
-    transition: background-color 0.25s, transform 0.1s; /* 배경색 및 변형 전환 효과 */
-}
-
-#check-email:hover {
-    color: #3c86bf; /* 글자 색상 */
-    border: #3c86bf solid 2px;
-}
-
-.form__span {
-    margin: 0; /* 상단 여백 제거 */
-    padding: 5px 0; /* 적당한 패딩 추가 */
-    font-size: 13px; /* 글자 크기 조정 */
-    line-height: 1.5; /* 줄 높이 조정 */
-    display: none; /* 기본적으로 숨김 */
-}
