--- templates/register.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Register</title>
-    
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            document.getElementById('check-username').addEventListener('click', function() {
-                const username = document.getElementById('username').value;
-                // 사용자 이름이 1글자 이상인지 확인
-                if (username.length < 1) {
-                    alert('Username must be at least 1 character long.');
-                    return;  // 조건을 만족하지 않으면 요청을 보내지 않음
-                }
-                fetch(`/check_username?username=${username}`)
-                    .then(response => response.json())
-                    .then(data => {
-                        const resultSpan = document.getElementById('username-check-result');
-                        if (data.exists) {
-                            resultSpan.textContent = 'Username already exists.';
-                            resultSpan.style.color = 'red';
-                        } else {
-                            resultSpan.textContent = 'Username is available.';
-                            resultSpan.style.color = 'green';
-                        }
-                    })
-                    .catch(error => {
-                        console.error('Error checking username:', error);
-                    });
-            });
-        });
-    </script>
-</head>
-<body>
-    <h2>Register</h2>
-    <form method="POST">
-        <label for="name">Name:</label>
-        <input type="text" id="name" name="name" required>
-
-        <br>
-        <label for="username">Username:</label>
-        <input type="text" id="username" name="username" required>
-        <button type="button" id="check-username">Check UserName</button>
-        <span id="username-check-result"></span>
-
-        <br>
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required>
-
-        <br>
-        <label for="sex">Sex:</label>
-        <select id="sex" name="sex" required>
-            <option value="m">Male</option>
-            <option value="f">Female</option>
-        </select>
-
-        <br>
-        <label for="age">Age:</label>
-        <input type="number" id="age" name="age" min="0" required>
-
-        <br>
-        <label for="email">Email:</label>
-        <input type="email" id="email" name="email" required>
-        <br>
-
-        <button type="submit">Register</button>
-    </form>
-    {% with messages = get_flashed_messages() %}
-        {% if messages %}
-            <ul>
-                {% for message in messages %}
-                    <li>{{ message }}</li>
-                {% endfor %}
-            </ul>
-        {% endif %}
-    {% endwith %}
-    <button onclick="window.location.href='/login'">로그인</button>
-</body>
-</html>
-<!-- <!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Login</title>
-    <link rel="stylesheet" href="/static/login_style.css">
-    <script>
-        function showAlert(message) {
-            alert(message);
-        }
-    </script>
-</head>
-<body>
-    <h2>Login</h2>
-    <form method="POST" id="loginForm">
-        <label for="username">Username:</label>
-        <input type="text" id="username" name="username" required>
-        <br>
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required>
-        <br>
-        <button type="submit">Login</button>
-    </form>
-    {% if username_error %}
-        <script>
-            showAlert("{{ username_error }}");
-        </script>
-    {% elif password_error %}
-        <script>
-            showAlert("{{ password_error }}");
-        </script>
-    {% endif %}
-    {% with messages = get_flashed_messages() %}
-        {% if messages %}
-            <ul>
-            {% for message in messages %}
-                <li>{{ message }}</li>
-            {% endfor %}
-            </ul>
-        {% endif %}
-    {% endwith %}
-
-    <button onclick="window.location.href='/register'">회원가입</button>
-    <script src="static/login_script.js"></script>
-</body>
-</html> -->
