+++ templates/type2.html
@@ -11,23 +11,6 @@
             // 페이지를 떠나기 전에 세션 초기화를 위한 요청을 보냅니다.
             navigator.sendBeacon('/reset_session');
         });
-        // Flash 메시지를 팝업으로 띄우는 함수
-        function showPopup(message) {
-        alert(message); // 간단한 팝업 메시지, 또는 모달 창을 사용할 수 있습니다.
-        }
-        function showPopup(message) {
-            alert(message); // 간단한 팝업 메시지, 또는 모달 창을 사용할 수 있습니다.
-        }
-
-        // 예를 들어, 페이지가 로드될 때 AJAX 요청을 보낸다고 가정해 봅시다.
-        fetch('/survey/type1')
-            .then(response => response.json())
-            .then(data => {
-                if (data.error) {
-                    showPopup(data.error);
-                }
-            })
-            .catch(error => console.error('Error:', error));
     </script>
     
 </head>
@@ -53,22 +36,15 @@
         <div id="chat-container">
             <div id="chat-box">
                 <div id="chat-header">
-                    <div id="chat-title">Type 1</div>
-                    <button id="reload-button1" class="reload-button">↺</button>
+                    <div id="chat-title">Type 2</div>
+                    <button id="reload-button2" class="reload-button">↺</button>
                 </div>
-                <div id="messages1" class="messages"></div>
+                <div id="messages2" class="messages"></div>
                 <div id="input-container">
-                    <textarea id="chat-input1" class="chat-input" placeholder="메시지를 입력하세요..."></textarea>
-                    <button id="send-button1" class="send-button">▲</button>
+                    <textarea id="chat-input2" class="chat-input" placeholder="메시지를 입력하세요..."></textarea>
+                    <button id="send-button2" class="send-button">▲</button>
                 </div>
             </div>
-        </div>
-    </div>
-    <div class="flash-messages">
-        <div class="flash-messages">
-            {% for category, message in messages %}
-                <p class="{{ category }}">{{ message }}</p>
-            {% endfor %}
         </div>
     </div>
     <script src="static/script.js"></script>
