--- test/KGtest.py
@@ -1,21 +1,0 @@
-from langchain_experimental.graph_transformers import LLMGraphTransformer
-from langchain_openai import ChatOpenAI
-from langchain_core.documents import Document
-
-import openai
-import os
-import dotenv
-
-dotenv.load_dotenv()
-openai.api_key = os.getenv('OPENAI_API_KEY')
-
-# Prompt used by LLMGraphTransformer is tuned for Gpt4.
-
-llm = ChatOpenAI(temperature=0, model="gpt-4")
-llm_transformer = LLMGraphTransformer(llm=llm)
-text = """ Marie Curie, was a Polish and naturalised-French physicist and chemist who conducted pioneering research on radioactivity. She was the first woman to win a Nobel Prize, the first person to win a Nobel Prize twice, and the only person to win a Nobel Prize in two scientific fields. Her husband, Pierre Curie, was a co-winner of her first Nobel Prize, making them the first-ever married couple to win the Nobel Prize and launching the Curie family legacy of five Nobel Prizes. She was, in 1906, the first woman to become a professor at the University of Paris. """
-documents = [Document(page_content=text)]
-
-graph_documents = llm_transformer.convert_to_graph_documents(documents)
-print(f"Nodes:{graph_documents[0].nodes}")
-print(f"Relationships:{graph_documents[0].relationships}")
